grpc user/requirements.txt
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
cryptography==36.0.1
dnspython==2.1.0
grpc-interceptor==0.13.1
grpcio==1.43.0
grpcio-tools==1.43.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itsdangerous==2.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
passlib==1.7.4
protobuf==3.19.3
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
swagger-ui-bundle==0.0.9
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==1.0.1
zipp==3.6.0
